Don't precreate all the columns for the torque_keys
Previous commit added support for automatically changing a torque_key to
populated when it receives data, and allows for having a torque_key
before the column is created. This commit creates the columns as
unpopulated, to clean up the GUI, and doesn't create the columns
before they are needed, which helps with space.